storage: Add I2C EEPROM backend, switch main to use it
Save to and read the config, calibration, user settings, from the
end of the MicroChip 24FC256 32kB onboard EEPROM memory over I2C.
On my device there were only 0s (or 0xff's, don't remember) in that area
before I overwrote it so we may be safe wrt. the SimpleBGC32 original
firmware's config and can coexist.

+++ src/storage.h
@@ -32,7 +32,7 @@ extern struct obgc_storage_config_s {
 #define MC_24FC256_BASE_ADDR 0x50
 
 void storage_init_internal_flash(void);
-void storage_init_i2c_eeprom(uint8_t addr, obgc_i2c *i2c, uint32_t size);
+void storage_init_i2c_eeprom(uint8_t addr, obgc_i2c *i2c, uint16_t size);
 
 int storage_read(void);
 int storage_write(void);
